+++ AddOns/Platynator/Locales.lua
@@ -257,6 +257,7 @@ L["NPC_ROLE"] = "NPC role"
 L["INCLUDE_TARGET"] = "Include target"
 L["ANIMATE"] = "Animate"
 L["ENABLE_IF_LINES_FALLING_OFF_FONT"] = "Enable if lines falling off font"
+L["EXECUTE"] = "Execute"
 
 L["WHAT_TO_EXPORT"] = "What to export?"
 L["STYLE"] = "Style"