+++ AddOns/Platynator/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Platynator
 
-## [293](https://github.com/TheMouseNest/Platynator/tree/293) (2026-01-29)
-[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/292...293) 
+## [295](https://github.com/TheMouseNest/Platynator/tree/295) (2026-01-30)
+[Full Changelog](https://github.com/TheMouseNest/Platynator/compare/294...295) 
 
-- Midnight Beta: Fix health being locked at 100% on health bar  
+- Add "Axe Toss" for Warlock interrupts  
+- Fix priority of paladin interrupt detection  
+- Fix Warlock interrupt, hopefully properly this time